+++ CHANGELOG.md
@@ -14,6 +14,9 @@ The format of this changelog is based on
 ## In progress
 
   - Added Apptainer/Singularity container build definition for Palace.
+  - Added build dependencies on [libCEED](https://github.com/CEED/libCEED) and
+    [LIBXSMM](https://github.com/libxsmm/libxsmm) to support operator partial assembly (CPU-
+    based for now).
 
 ## [0.11.2] - 2023-07-14